Bug 1478724: Upgrage broken due to multirecipientnotifications changes
When upgrading a site that existed before the plugin was added to core

behatnotneeded

+++ htdocs/lib/db/upgrade.php
@@ -3798,7 +3798,7 @@ function xmldb_core_upgrade($oldversion=0) {
 
     if ($oldversion < 2014092300) {
         log_debug("Install 'multirecipientnotification' plugin");
-        if ($data = check_upgrades('artefact.multirecipientnotification')) {
+        if ($data = check_upgrades('module.multirecipientnotification')) {
             upgrade_plugin($data);
         }
     }
@@ -4103,37 +4103,41 @@ function xmldb_core_upgrade($oldversion=0) {
     }
 
     if ($oldversion < 2015072000) {
-        log_debug('Change installation of artefact plugin multirecipentNotification to plugin module');
+        // If we are upgrading from a site built before 2014092300 straight to 15.10
+        // then the plugin won't exist as an artefact.
+        if (table_exists(new XMLDBTable('artefact_multirecipient_userrelation'))) {
+            log_debug('Change installation of artefact plugin multirecipentNotification to plugin module.');
 
-        // first, drop the old triggers
-        db_drop_trigger('update_unread_insert2', 'artefact_multirecipient_userrelation');
-        db_drop_trigger('update_unread_update2', 'artefact_multirecipient_userrelation');
-        db_drop_trigger('update_unread_delete2', 'artefact_multirecipient_userrelation');
+            // first, drop the old triggers
+            db_drop_trigger('update_unread_insert2', 'artefact_multirecipient_userrelation');
+            db_drop_trigger('update_unread_update2', 'artefact_multirecipient_userrelation');
+            db_drop_trigger('update_unread_delete2', 'artefact_multirecipient_userrelation');
 
-        // rename tables artefact_multirecipientnotifiaction_notification and
-        // Table: artefact_multirecipient_userrelation to module-prefix
-        execute_sql("ALTER TABLE {artefact_multirecipient_notification} RENAME TO {module_multirecipient_notification}");
-        execute_sql("ALTER TABLE {artefact_multirecipient_userrelation} RENAME TO {module_multirecipient_userrelation}");
+            // rename tables artefact_multirecipientnotifiaction_notification and
+            // Table: artefact_multirecipient_userrelation to module-prefix
+            execute_sql("ALTER TABLE {artefact_multirecipient_notification} RENAME TO {module_multirecipient_notification}");
+            execute_sql("ALTER TABLE {artefact_multirecipient_userrelation} RENAME TO {module_multirecipient_userrelation}");
 
-        //move event_subscrition entries for artefact plugin
-        //multirecipientnotification to table module_event_subscription
-        $subscriptions = get_records_array('artefact_event_subscription', 'plugin', 'multirecipientnotification');
-        delete_records('artefact_event_subscription', 'plugin', 'multirecipientnotification');
-        delete_records('artefact_installed_type', 'plugin', 'multirecipientnotification');
-        $installrecord = get_record('artefact_installed', 'name', 'multirecipientnotification');
-        if (is_object($installrecord)) {
-            insert_record('module_installed', $installrecord);
-            delete_records('artefact_installed', 'name', 'multirecipientnotification');
-        }
-        if (is_array($subscriptions)) {
-            foreach ($subscriptions as $subscription) {
-                insert_record('module_event_subscription', $subscription, 'id');
+            //move event_subscrition entries for artefact plugin
+            //multirecipientnotification to table module_event_subscription
+            $subscriptions = get_records_array('artefact_event_subscription', 'plugin', 'multirecipientnotification');
+            delete_records('artefact_event_subscription', 'plugin', 'multirecipientnotification');
+            delete_records('artefact_installed_type', 'plugin', 'multirecipientnotification');
+            $installrecord = get_record('artefact_installed', 'name', 'multirecipientnotification');
+            if (is_object($installrecord)) {
+                insert_record('module_installed', $installrecord);
+                delete_records('artefact_installed', 'name', 'multirecipientnotification');
             }
-        }
+            if (is_array($subscriptions)) {
+                foreach ($subscriptions as $subscription) {
+                    insert_record('module_event_subscription', $subscription, 'id');
+                }
+            }
 
-        // recreate trigger
-        safe_require('module', 'multirecipientnotification');
-        PluginModuleMultirecipientnotification::postinst(0);
+            // recreate trigger
+            safe_require('module', 'multirecipientnotification');
+            PluginModuleMultirecipientnotification::postinst(0);
+        }
     }
 
     return $status;
